+++ Tests/Bower/Fixtures/simple_dependency_mapping.php
@@ -1,0 +1,12 @@
+<?php
+return array(
+    "simple_package" => array(
+        "source" => array(
+            "main" => array(
+                "../components/simple_package/styles.css",
+                "../components/simple_package/script.js",
+                "../components/simple_package/image.png"
+            ),
+        ),
+    )
+);