+++ Changelog.md
@@ -3,7 +3,9 @@ Changelog
 
 ### 0.6-dev
 
-...
+* When the retrieved mapping from the bower command ```bower list --map``` is invalid, the dependency cache will be deleted
+* The dependency cache warmer no longer creates the cache twice when install_on_warmup is enabled
+* An exception is now thrown when a required js/css file could not be found when retrieving the dependency mapping
 
 ### 0.5 (2013-03-14)